# Avoid some virtual dispatch in Encodings