- Seal some internal overrides
- In the more common encodings, add overrides of GetBytes/GetByteCount/GetChars/GetCharCount to avoid another layer of virtuals from the base class (they can call to an internal non-virtual).

Related to https://github.com/dotnet/corefx/issues/30382

cc: @jkotas, @roji, @tarekgh, @krwq