@stephentoub 2018-06-15
Example microbenchmark:
```C#
using System;
using System.Diagnostics;
using System.Text;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using BenchmarkDotNet.Attributes.Jobs;

[MemoryDiagnoser]
[InProcess]
public class Benchmarks
{
    private static void Main() => BenchmarkRunner.Run<Benchmarks>();

    private char[] _chars = "hello world".ToCharArray();
    private byte[] _bytes = new byte[100];

    [Benchmark]
    public void EncodingUTF8GetBytes()
    {
        ReadOnlySpan<char> src = _chars;
        Span<byte> dst = _bytes;
        for (int i = 0; i < 1000; i++)
        {
            Encoding.UTF8.GetBytes(src, dst);
        }
    }
}
```

|               Method |     Mean |     Error |    StdDev | Allocated |
--------------------- |---------:|----------:|----------:|----------:|
| Before | 21.43 us | 0.1606 us | 0.1502 us |       0 B |
| After | 19.53 us | 0.1589 us | 0.1409 us |       0 B |